SUBS-320 Other MG Amount
* Added other amount currency input and validation for custom MG Amount
* Added validation to 8500 to allow room for 15% donation
* rearranged social share above receipt
* Cleaned up old MG form and route
