+++ src/assets/scss/global/forms.scss
@@ -55,6 +55,12 @@ label.error {
 	}
 }
 
+// for inputs without a label
+input.error {
+	border: 1px solid $kiva-accent-red;
+}
+
+// validation error text
 .validation-errors {
 	margin-left: 0;
 }
